requirements.txt
viam-sdk==0.28.0
google-api-python-client==2.0.0
pyinstaller==6.3.0
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy==2.1.0